airflow/requirements.txt
apache-airflow==2.7.3
pandas==2.1.3
boto3==1.29.7
psycopg2-binary==2.9.9

